tauri/python-parking-system/requirements.txt
# Python Parking System Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
aiosqlite==0.19.0
fast-alpr==0.2.0
opencv-python==4.9.0.80
pillow==10.1.0
numpy==1.24.3
httpx==0.25.2
websockets==12.0
python-multipart==0.0.6
jinja2==3.1.2
python-socketio==5.10.0
eventlet==0.33.3
pyserial==3.5
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
typer==0.9.0
rich==13.7.0
watchdog==3.0.0
requests==2.31.0
pygame==2.5.2
couchdb==1.2
typing-extensions==4.8.0
python-dotenv==1.0.0
